requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr!=2.1.0,>=2.0.0 # Apache-2.0

# The greenlet package must be compiled with gcc and needs
# the Python.h headers. Make sure you install the python-dev
# package to get the right headers...
greenlet>=0.4.10 # MIT

eventlet!=0.18.3,!=0.20.1,>=0.18.2 # MIT
PasteDeploy>=1.5.0 # MIT
Routes>=2.3.1 # MIT
WebOb>=1.8.4 # MIT
oslo.config>=5.2.0 # Apache-2.0
oslo.concurrency>=3.26.0 # Apache-2.0
oslo.context>=2.19.2 # Apache-2.0
oslo.middleware>=3.31.0 # Apache-2.0
oslo.service>=1.33.0 # Apache-2.0
oslo.upgradecheck>=0.1.0 # Apache-2.0
oslo.utils>=3.33.0 # Apache-2.0
stevedore>=1.20.0 # Apache-2.0
keystonemiddleware>=4.17.0 # Apache-2.0
WSME>=0.8.0 # MIT

# For paste.util.template used in keystone.common.template
Paste>=2.0.2 # MIT

python-keystoneclient>=3.8.0 # Apache-2.0
pyOpenSSL>=17.1.0 # Apache-2.0
# Required by openstack.common libraries

oslo.i18n>=3.15.3 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.reports>=1.18.0 # Apache-2.0
oslo.messaging>=5.29.0 # Apache-2.0
oslo.policy>=1.30.0 # Apache-2.0
oslo.serialization!=2.19.1,>=2.18.0 # Apache-2.0

osprofiler>=1.4.0 # Apache-2.0

# ElasticSearch 6.x
# elasticsearch>=6.0.0,<7.0.0
# ElasticSearch 5.x
# elasticsearch>=5.0.0,<6.0.0
# ElasticSearch 2.x
# elasticsearch>=2.0.0,<3.0.0
# See:
# https://elasticsearch-py.readthedocs.io/en/master/#compatibility
elasticsearch<3.0.0,>=2.0.0 # Apache-2.0

# Clients required for indexing plugins
python-designateclient>=2.7.0 # Apache-2.0
python-glanceclient>=2.8.0 # Apache-2.0
python-novaclient>=9.1.0 # Apache-2.0
python-neutronclient>=6.11.0 # Apache-2.0
python-cinderclient>=3.3.0 # Apache-2.0
python-swiftclient>=3.2.0 # Apache-2.0
python-ironicclient>=2.3.0 # Apache-2.0
